requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-material==1.7.0
django-registration==3.1
django-tinymce==3.0.2
djangorestframework==3.11.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.1.0
